changed to single precision floating point

+++ src/e_flop.c
@@ -9,33 +9,27 @@ int main(void)
 	long n = INTERVALS;
 	long i;
 	unsigned my_row, my_col;
-	double mypi, h, sum, x;
-	double *partial = (void *)0x6000;
+	float mypi, h, sum, x;
+	float *partial = (void *)0x6000;
 	unsigned *done = (void *)0x7000;
 
-	// partial = (unsigned *) 0x6000;
-	// done = (unsigned *) 0x7000;
-
-	// (*(partial)) = 0.0;
-	(*partial) = 0.0;
+	(*partial) = (float)0.0;
 
 	coreid = e_get_coreid();
 	e_coords_from_coreid(coreid, &my_row, &my_col);
 
-	h = 1.0 / (double) n;
-	sum = 0.0;
+	h = (float)1.0 / (float) n;
+	sum = (float)0.0;
 
 	for (i = (my_row * 4) + (my_col + 1); i <= n; i += CORES)
 	{
-		x = h * ((double)i - 0.5);
-		sum += ( 4.0 / (1.0 + x * x ));
+		x = h * ((float)i - (float)0.5);
+		sum += ( (float)4.0 / ((float)1.0 + x * x ));
 	}
 	mypi = h * sum;
 
-	// (*(partial)) = mypi;
 	(*partial) = mypi;
 
-	// (*(done)) = 0x00000001;
 	(*done) = 0x00000001;
 
 	__asm__ __volatile__("idle");
